Move readValue() to ValueToObjectConverter(2)
