+++ h2/src/main/org/h2/value/DataType.java
@@ -5,41 +5,20 @@
  */
 package org.h2.value;
 
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.Reader;
-import java.math.BigDecimal;
-import java.nio.charset.StandardCharsets;
-import java.sql.Array;
-import java.sql.Date;
 import java.sql.JDBCType;
-import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.SQLType;
-import java.sql.Time;
-import java.sql.Timestamp;
 import java.sql.Types;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetDateTime;
-import java.time.OffsetTime;
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.UUID;
 
 import org.h2.api.ErrorCode;
 import org.h2.api.H2Type;
-import org.h2.api.Interval;
 import org.h2.api.IntervalQualifier;
 import org.h2.engine.Mode;
-import org.h2.engine.SessionInterface;
 import org.h2.engine.SysProperties;
 import org.h2.message.DbException;
-import org.h2.util.JSR310Utils;
-import org.h2.util.LegacyDateTimeUtils;
 import org.h2.util.StringUtils;
 
 /**
@@ -484,309 +463,6 @@ public class DataType {
      */
     public static ArrayList<DataType> getTypes() {
         return TYPES;
-    }
-
-    /**
-     * Read a value from the given result set.
-     *
-     * @param session the session
-     * @param rs the result set
-     * @param columnIndex the column index (1 based)
-     * @param type the data type
-     * @return the value
-     */
-    public static Value readValue(SessionInterface session, ResultSet rs, int columnIndex, int type) {
-        try {
-            Value v;
-            switch (type) {
-            case Value.NULL:
-                v = ValueNull.INSTANCE;
-                break;
-            case Value.VARBINARY: {
-                /*
-                 * Both BINARY and UUID may be mapped to Value.BYTES. getObject() returns byte[]
-                 * for SQL BINARY, UUID for SQL UUID and null for SQL NULL.
-                 */
-                Object o = rs.getObject(columnIndex);
-                if (o instanceof byte[]) {
-                    v = ValueVarbinary.getNoCopy((byte[]) o);
-                } else if (o != null) {
-                    v = ValueUuid.get((UUID) o);
-                } else {
-                    v = ValueNull.INSTANCE;
-                }
-                break;
-            }
-            case Value.BINARY: {
-                byte[] bytes = rs.getBytes(columnIndex);
-                if (bytes != null) {
-                    v = ValueBinary.getNoCopy(bytes);
-                } else {
-                    v = ValueNull.INSTANCE;
-                }
-                break;
-            }
-            case Value.UUID: {
-                Object o = rs.getObject(columnIndex);
-                if (o instanceof UUID) {
-                    v = ValueUuid.get((UUID) o);
-                } else if (o != null) {
-                    v = ValueUuid.get((byte[]) o);
-                } else {
-                    v = ValueNull.INSTANCE;
-                }
-                break;
-            }
-            case Value.BOOLEAN: {
-                boolean value = rs.getBoolean(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueBoolean.get(value);
-                break;
-            }
-            case Value.TINYINT: {
-                byte value = rs.getByte(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueTinyint.get(value);
-                break;
-            }
-            case Value.DATE: {
-                try {
-                    Object value = rs.getObject(columnIndex, LocalDate.class);
-                    v = value == null ? ValueNull.INSTANCE : JSR310Utils.localDateToValue(value);
-                    break;
-                } catch (SQLException ignore) {
-                    // Nothing to do
-                }
-                Date value = rs.getDate(columnIndex);
-                v = value == null ? ValueNull.INSTANCE : LegacyDateTimeUtils.fromDate(session, null, value);
-                break;
-            }
-            case Value.TIME: {
-                try {
-                    Object value = rs.getObject(columnIndex, LocalTime.class);
-                    v = value == null ? ValueNull.INSTANCE : JSR310Utils.localTimeToValue(value);
-                    break;
-                } catch (SQLException ignore) {
-                    // Nothing to do
-                }
-                Time value = rs.getTime(columnIndex);
-                v = value == null ? ValueNull.INSTANCE : LegacyDateTimeUtils.fromTime(session, null, value);
-                break;
-            }
-            case Value.TIME_TZ: {
-                try {
-                    Object value = rs.getObject(columnIndex, OffsetTime.class);
-                    v = value == null ? ValueNull.INSTANCE : JSR310Utils.offsetTimeToValue(value);
-                    break;
-                } catch (SQLException ignore) {
-                    // Nothing to do
-                }
-                Object obj = rs.getObject(columnIndex);
-                if (obj == null) {
-                    v = ValueNull.INSTANCE;
-                } else {
-                    v = ValueTimeTimeZone.parse(obj.toString());
-                }
-                break;
-            }
-            case Value.TIMESTAMP: {
-                try {
-                    Object value = rs.getObject(columnIndex, LocalDateTime.class);
-                    v = value == null ? ValueNull.INSTANCE : JSR310Utils.localDateTimeToValue(value);
-                    break;
-                } catch (SQLException ignore) {
-                    // Nothing to do
-                }
-                Timestamp value = rs.getTimestamp(columnIndex);
-                v = value == null ? ValueNull.INSTANCE : LegacyDateTimeUtils.fromTimestamp(session, null, value);
-                break;
-            }
-            case Value.TIMESTAMP_TZ: {
-                try {
-                    Object value = rs.getObject(columnIndex, OffsetDateTime.class);
-                    v = value == null ? ValueNull.INSTANCE : JSR310Utils.offsetDateTimeToValue(value);
-                    break;
-                } catch (SQLException ignore) {
-                    // Nothing to do
-                }
-                Object obj = rs.getObject(columnIndex);
-                if (obj == null) {
-                    v = ValueNull.INSTANCE;
-                } else if (obj instanceof ZonedDateTime) {
-                    v = JSR310Utils.zonedDateTimeToValue(obj);
-                } else {
-                    v = ValueTimestampTimeZone.parse(obj.toString(), session);
-                }
-                break;
-            }
-            case Value.NUMERIC: {
-                BigDecimal value = rs.getBigDecimal(columnIndex);
-                v = value == null ? ValueNull.INSTANCE : ValueNumeric.get(value);
-                break;
-            }
-            case Value.DOUBLE: {
-                double value = rs.getDouble(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueDouble.get(value);
-                break;
-            }
-            case Value.REAL: {
-                float value = rs.getFloat(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueReal.get(value);
-                break;
-            }
-            case Value.INTEGER: {
-                int value = rs.getInt(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueInteger.get(value);
-                break;
-            }
-            case Value.BIGINT: {
-                long value = rs.getLong(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueBigint.get(value);
-                break;
-            }
-            case Value.SMALLINT: {
-                short value = rs.getShort(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueSmallint.get(value);
-                break;
-            }
-            case Value.VARCHAR_IGNORECASE: {
-                String s = rs.getString(columnIndex);
-                v = (s == null) ? ValueNull.INSTANCE : ValueVarcharIgnoreCase.get(s);
-                break;
-            }
-            case Value.CHAR: {
-                String s = rs.getString(columnIndex);
-                v = (s == null) ? ValueNull.INSTANCE : ValueChar.get(s);
-                break;
-            }
-            case Value.VARCHAR: {
-                String s = rs.getString(columnIndex);
-                v = (s == null) ? ValueNull.INSTANCE : ValueVarchar.get(s);
-                break;
-            }
-            case Value.CLOB: {
-                if (session == null) {
-                    String s = rs.getString(columnIndex);
-                    v = s == null ? ValueNull.INSTANCE :
-                        ValueLobInMemory.createSmallLob(Value.CLOB, s.getBytes(StandardCharsets.UTF_8));
-                } else {
-                    Reader in = rs.getCharacterStream(columnIndex);
-                    if (in == null) {
-                        v = ValueNull.INSTANCE;
-                    } else {
-                        v = session.addTemporaryLob(
-                                session.getDataHandler().getLobStorage().createClob(new BufferedReader(in), -1));
-                    }
-                }
-                break;
-            }
-            case Value.BLOB: {
-                if (session == null) {
-                    byte[] buff = rs.getBytes(columnIndex);
-                    return buff == null ? ValueNull.INSTANCE : ValueLobInMemory.createSmallLob(Value.BLOB, buff);
-                }
-                InputStream in = rs.getBinaryStream(columnIndex);
-                if (in == null) {
-                    v = ValueNull.INSTANCE;
-                } else {
-                    v = session.addTemporaryLob(session.getDataHandler().getLobStorage().createBlob(in, -1));
-                }
-                break;
-            }
-            case Value.JAVA_OBJECT: {
-                byte[] buff = rs.getBytes(columnIndex);
-                v = buff == null ? ValueNull.INSTANCE : ValueJavaObject.getNoCopy(buff);
-                break;
-            }
-            case Value.ARRAY: {
-                Array array = rs.getArray(columnIndex);
-                if (array == null) {
-                    return ValueNull.INSTANCE;
-                }
-                Object[] list = (Object[]) array.getArray();
-                if (list == null) {
-                    return ValueNull.INSTANCE;
-                }
-                int len = list.length;
-                Value[] values = new Value[len];
-                for (int i = 0; i < len; i++) {
-                    values[i] = ValueToObjectConverter.objectToValue(session, list[i], Value.NULL);
-                }
-                v = ValueArray.get(values, session);
-                break;
-            }
-            case Value.ENUM: {
-                int value = rs.getInt(columnIndex);
-                v = rs.wasNull() ? ValueNull.INSTANCE : ValueInteger.get(value);
-                break;
-            }
-            case Value.ROW: {
-                Object[] list = (Object[]) rs.getObject(columnIndex);
-                if (list == null) {
-                    return ValueNull.INSTANCE;
-                }
-                int len = list.length;
-                Value[] values = new Value[len];
-                for (int i = 0; i < len; i++) {
-                    values[i] = ValueToObjectConverter.objectToValue(session, list[i], Value.NULL);
-                }
-                v = ValueRow.get(values);
-                break;
-            }
-            case Value.RESULT_SET: {
-                ResultSet x = (ResultSet) rs.getObject(columnIndex);
-                if (x == null) {
-                    return ValueNull.INSTANCE;
-                }
-                return ValueResultSet.get(session, x, Integer.MAX_VALUE);
-            }
-            case Value.GEOMETRY: {
-                Object x = rs.getObject(columnIndex);
-                if (x == null) {
-                    return ValueNull.INSTANCE;
-                }
-                return ValueGeometry.getFromGeometry(x);
-            }
-            case Value.INTERVAL_YEAR:
-            case Value.INTERVAL_MONTH:
-            case Value.INTERVAL_DAY:
-            case Value.INTERVAL_HOUR:
-            case Value.INTERVAL_MINUTE:
-            case Value.INTERVAL_SECOND:
-            case Value.INTERVAL_YEAR_TO_MONTH:
-            case Value.INTERVAL_DAY_TO_HOUR:
-            case Value.INTERVAL_DAY_TO_MINUTE:
-            case Value.INTERVAL_DAY_TO_SECOND:
-            case Value.INTERVAL_HOUR_TO_MINUTE:
-            case Value.INTERVAL_HOUR_TO_SECOND:
-            case Value.INTERVAL_MINUTE_TO_SECOND: {
-                Object x = rs.getObject(columnIndex);
-                if (x == null) {
-                    return ValueNull.INSTANCE;
-                }
-                Interval interval = (Interval) x;
-                return ValueInterval.from(interval.getQualifier(), interval.isNegative(),
-                        interval.getLeading(), interval.getRemaining());
-            }
-            case Value.JSON: {
-                Object x = rs.getObject(columnIndex);
-                if (x == null) {
-                    return ValueNull.INSTANCE;
-                }
-                Class<?> clazz = x.getClass();
-                if (clazz == byte[].class) {
-                    return ValueJson.fromJson((byte[]) x);
-                } else if (clazz == String.class) {
-                    return ValueJson.fromJson((String) x);
-                } else {
-                    return ValueJson.fromJson(x.toString());
-                }
-            }
-            default:
-                throw DbException.throwInternalError("type="+type);
-            }
-            return v;
-        } catch (SQLException e) {
-            throw DbException.convert(e);
-        }
     }
 
     /**
